requirements.txt
streamlit
pandas
langchain
langgraph
langchain_community
langchain_groq
sentence-transformers
faiss-cpu